Remove `Oracle` type
Oracle has never really been a value type, just something that can be
named with an identifier. Therefore, it shouldn't be in the Type enum.
We instead added a "ScopeType" enum that only needs to be used
in that particular submodule, that is either a proper type or an oracle.

We have a similar problem with Fn, but it's different and needs a
different solution, so we didn't address it here.