remove superfluous (and wrong) check
we move the check to the parser. also, we can't simply compare the
packages, since the constant identifier expressions it contains have the
name of the game in there.

In the parser we now check that the package name (not instance name! the
type so to speak) as well as all paramters are the same.