Refactor Sorts so we can easier generate relation functions
This commit is huge. The bigger changes are listed below, but there also
have been some drive-by cleanup.

Preparation and progress toward generating relation functions
-------------------------------------------------------------

We haven't been generating a lot of functions except for oracles, and
this doing this by hand has been pretty annoying. One problem is the
complexity introduced by the SmtSort trait, where each Sort would be
it's own type. It requires a lot of type annotations and prevents us
from building e.g. a vector where the elements contain a sort.

Refactor Sorts
--------------

Refactors sort handling from trait-based polymorphism to a single Sort
enum. While it isn't as strong of a model when it comes to the types, it
is a lot simpler. I even tried building SmtSort trait objects, but that
also didn't work (maybe because I was holding it wrong).

I also ported the entire code to the new code and removed the old
SmtSorts. It looks like the approach works well, but it might make sense
to revisit what variants the Sort enum should really have. Should there
be a separate thing for datatypes? If so, why? What does the SMT-LIB
spec say about what a sort is, syntactically?

Better handle parametric datatypes
----------------------------------

Made some changes to the DatastructurePattern to better handle
parametric types such as Maybe. Side-effect by the Sorts refactor.